# openapi/schema.yaml
openapi: 3.0.3
info:
  title: Merchant Crypto Wallet Dashboard API
  version: 0.1.0
  description: |
    MVP API contract for merchant crypto wallet dashboard

    ## Authentication
    This API uses Supabase Auth for authentication. Users authenticate via OTP (One-Time Password)
    sent to their email. The authentication flow is handled directly by Supabase Auth client,
    not through these API endpoints.

    For protected endpoints, include the Supabase session access token in the Authorization header:
    `Authorization: Bearer <supabase_access_token>`
servers:
  - url: /api
    description: API base URL
paths:
...
  /wallet/send:
    post:
      summary: Send MNEE to another wallet
      description: |
        Create an outgoing transaction for sending MNEE to another wallet address.
        This will:
        - **Validate sufficient balance** before allowing the transaction
        - Create a transaction with type='out', method='transfer', display_type='send', status='posted'
        - Store recipient address in customer_address field
        - Generate mock blockchain transaction hash (send_hash)
        - Assign to default "Wallet" module
        - Return real transaction ID, transaction details, and mock blockchain hash
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
                - recipient
                - address
              properties:
                amount:
                  type: number
                  format: double
                  description: Amount of MNEE to send
                recipient:
                  type: string
                  description: Name or identifier of the recipient
                address:
                  type: string
                  description: Wallet address of the recipient (Ethereum format)
      responses:
        '201':
          description: Transaction created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  transactionId:
                    type: string
                  transaction:
                    $ref: '#/components/schemas/Transaction'
                  txHash:
                    type: string
                    description: Mock blockchain transaction hash (0x + 64 hex chars)
        '400':
          description: Invalid request data or insufficient balance
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  details:
                    type: object
                    properties:
                      requested:
                        type: number
                      available:
                        type: number
        '401':
          description: Unauthorized
  /wallet/receive:
    post:
      summary: Generate wallet address and QR code for receiving MNEE
      description: |
        Generate a mock wallet address and QR code for receiving MNEE.
        This will:
        - Generate mock Ethereum wallet address (0x + 40 hex characters)
        - Generate QR code as base64 data URL using the address
        - If amount is provided, encode it in the QR code for payment request
        - **Does NOT create a transaction** (transaction created when funds are actually received)
        - Return wallet address and QR code data
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  format: double
                  description: Optional amount for payment request QR code
      responses:
        '200':
          description: Wallet address and QR code generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  address:
                    type: string
                    description: Mock Ethereum wallet address (0x + 40 hex chars)
                  qrCode:
                    type: string
                    description: Base64 encoded QR code data URL
                  amount:
                    type: number
                    description: Amount if provided in request
        '400':
          description: Invalid request data
        '401':
          description: Unauthorized
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TransactionType:
      type: string
      enum:
        - in
        - out
    TransactionStatus:
      type: string
      enum:
        - pending
        - posted
        - failed
    Transaction:
      type: object
      required:
        - id
        - moduleId
        - type
        - method
        - displayType
        - amount
        - currency
        - status
        - createdAt
      properties:
        id:
          type: string
        moduleId:
          type: string
        type:
          $ref: '#/components/schemas/TransactionType'
        method:
          type: string
          enum:
            - stablecoin
            - payout
            - adjustment
            - card
            - bank
            - transfer
        displayType:
          type: string
          enum:
            - send
            - receive
            - buy
            - sell
            - swap
        amount:
          type: number
          format: double
        currency:
          type: string
        status:
          $ref: '#/components/schemas/TransactionStatus'
        createdAt:
          type: string
          format: date-time
        customerEmail:
          type: string
        sendHash:
          type: string
        rockWalletId:
          type: string
        txHashIn:
          type: string
        txHashSwap:
          type: string
        feeUsd:
          type: number
          format: double
        customerAddress:
          type: string
...