# src/lib/api/generated/schemas/transaction.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * Merchant Crypto Wallet Dashboard API
 * MVP API contract for merchant crypto wallet dashboard

## Authentication
This API uses Supabase Auth for authentication. Users authenticate via OTP (One-Time Password)
sent to their email. The authentication flow is handled directly by Supabase Auth client,
not through these API endpoints.

For protected endpoints, include the Supabase session access token in the Authorization header:
`Authorization: Bearer <supabase_access_token>`

 * OpenAPI spec version: 0.1.0
 */
import type { TransactionType } from "./transactionType";
import type { TransactionMethod } from "./transactionMethod";
import type { TransactionDisplayType } from "./transactionDisplayType";
import type { TransactionStatus } from "./transactionStatus";

export interface Transaction {
  id: string;
  moduleId: string;
  type: TransactionType;
  method: TransactionMethod;
  displayType: TransactionDisplayType;
  amount: number;
  currency: string;
  status: TransactionStatus;
  createdAt: string;
  customerEmail?: string;
  sendHash?: string;
  rockWalletId?: string;
  txHashIn?: string;
  txHashSwap?: string;
  feeUsd?: number;
  customerAddress?: string;
}